Broken attempt at using portal to refactor CrossingDetail
