+++ frontend/package.json
@@ -62,5 +62,6 @@
   },
   "eslintConfig": {
     "extends": "react-app"
-  }
+  },
+  "moduleRoots": ["src"]
 }
